Extract internal/common/testing/container to internal/containertest
