+++ internal/containertest/container_integration_test.go
@@ -15,7 +15,7 @@
 //go:build integration
 // +build integration
 
-package container
+package containertest
 
 import (
 	"context"
